Add support for L?[x] list access syntax (#262)
I renamed the current `IndexKind` enum to `PropertyAccessKind` purely 
based off of the fact that somewhere in the byond ref both procs and 
vars are referred to as a datum's "properties" and I had no better name 
ideas.

The checker treats conditional list accesses exactly the same as normal 
ones. I was going to add some linting stuff for them but their expected 
behaviour is completely up in the air at the moment.