auxtools debug server: expression evaluation (#263)
The changes for the language server are minimal. This will require 
shipping version 2.0.0 of the debug server: use 
https://github.com/willox/auxtools/releases/tag/v2.2.0

There are plenty of compiler features missing on the debug server's 
end, but it is definitely more than usable so I want to get it 
published. I plan on going a lot further this all of this, maybe we'll 
even have hot loading one day.

Noticeable differences from actual BYOND are:
1. It only compiles expressions! This means no for loops, no if 
   statements, etc.
2. It lacks all static typing - the compiler has no access to type 
   definitions and will treat all `.` accesses as `:` accesses and 
   ignore the rules of const.
3. There are no implicit uses of `src`. This comes from the fact that 
   there is no static typing, so to access a field on `src` you have to 
   explicitly use `src.xyz` syntax.
4. Strings are lame. There is no interpolation support and the escape 
   sequences are super limited too.
5. Some built-in procs aren't supported, but most are. They'll 
   gracefully error if you try to use them.
6. No pre-processor.

For other limitations, just view the error list: 
https://github.com/willox/dmasm/blob/9ba328ef8c4e45c64072198e72c412cd9f685609/src/compiler.rs#L45-L76

tl;dr: relatively well-working debug console eval, watch eval, and 
conditional breakpoints