Add hide_invisible config and only-wires-and-pipes pass (#272)
This PR adds in a configuration for the `hide_invisible` render pass, 
to add extra types to it to forcibly hide. Configuration is as follows:
```toml
[map_renderer]
hide_invisible = [
    "/obj/effect/mapping_helpers",
    "/obj/effect/landmark",
    "/obj/effect/spawner/random_spawners",
]
```

This PR also adds a new render pass, `only-wires-and-pipes`. This is 
functionally a combination of `only-wires` and `only-pipes`, as you 
currently cannot stack both the passes.

Co-authored-by: Tad Hardesty <tad@platymuus.com>