Add [u8] read/write functions to Image/IconFile/Metadata
This is so IconDiffBot2 can operate purely in memory.

Metadata::from_raw and IconFile::from_raw take an AsRef<[u8]>
as the sequence of bytes of a valid PNG+DMI file to produce their types.

Image::to_raw gives a Vec<u8> which is just the output from
png::Encoder.

There is no Image::from_raw because there's no reason not to use
IconFile::from_raw, and there is no to_raw for Metadata or IconFile
because they don't support to_file either. IconFile could, but it's
unnecessary.