Add /dm_filter hidden type to builtins (#312)
reference to this type -
<https://www.byond.com/forum/post/2544111#comment25180960>

After some cursory poking, /dm_filter is a hidden type that can be used to manipulate filter instances without using the runtime search operator (:). There is no official reference material for it. It does not descend from datum, cannot be subtyped, and can only be created *successfully* by a valid call to proc/filter(...).  All filter types create the same kind of /dm_filter but with different properties, of which I believe type is the only const.

A contributor on Bay tried to use it but checker doesn't know about it. I'm not familiar with the guts of suite, but I think this is all that's necessary?