Implements /proc/final, piped into should not override
As a part of this we need to make a struct when building proc defs, to
keep track of their flags AND their kind at the same time.

We then pass that down into a new flags var on the proc decl struct, and
we're golden

Only thing of note is I removed the is_private and is_protected vars
from proc declerations because they were totally unused. Spooky added em
a long time ago and I think it was for not much reason