Implements compile time warning for datum[]
Lummox recently made this a RUNTIME error, but also irrrespective of
that this tends to mask HELL vars[] security holes. We really should be
linting for it.

I had to implmenet operator overloading support for the builtin parser,
which was far more annoying then I expected. Bit scuffed, but it works.
(I am unsure of how to parse for "" though so I left that unimplemented)

This will likely throw errors on most codebases, I'm unsure of how to
handle that..