Fix preprocessor relative filepath evaluation for fexists (#401)
The way I implemented const-eval'd fexists is slightly incorrect. It's 
evaulating the existence of the path relative to the working directory, 
instead of relative to the current file. This is causing errors in 
StrongDMM, because it's checking paths relative to the application exe 
file.

I tested by building StrongDMM and using this fixed version, and it 
worked correctly.