Add a body_range field to ProcValue (#438)
I wasn't able to find a concise way to reference the range and/or
contents of proc bodies without also needing to reference the Annotation
tree. I haven't implemented any use case for the added information
outside of some experiments I'm doing with the deeper analysis
companion; I think it could be useful for determining return values and
types without the need for SpacemanDMM directives, at a later point.

It's an Option to account for built-ins not having any actual presence
in the syntax tree.